Setup tests for the new work
