+++ demo/filters-001.html
@@ -58,88 +58,15 @@
     <a href="filters-002.html">Next &rarr;</a>
   </div>
   <h1><a href="index.html">Scrawl-canvas v8</a> - Filters test 001</h1>
-  <h2>Parameters for: Blur, Gaussianblur filters; filter memoization</h2>
+  <h2>Parameters for: red, green, blue, cyan, magenta, yellow, notred, notgreen, notblue, grayscale, sepia, invert filters</h2>
   <h3>To change image, drag-and-drop an image file over the canvas</h3>
 
   <div class="controls">
     
-    <div class="lavender label">Blur</div>
-    <div class="lavender">
-      <select class="controlItem" id="blurFilter">
-        <option value="legacy-blur">Legacy Blur</option>
-        <option value="gaussian-blur">Gaussian blur</option>
-      </select>
-    </div>
-
     <div class="lavender label">Opacity</div>
     <div class="lavender"><input class="controlItem" id="opacity" type="range" value="1" min="0" max="1" step="0.01" /></div>
 
-    <div class="yellow label">Radius</div>
-    <div class="yellow"><input class="controlItem" id="radius" type="range" value="10" min="0" max="60" step="1" /></div>
-
-    <div class="yellow label">Passes</div>
-    <div class="yellow"><input class="controlItem" id="passes" type="range" value="1" min="0" max="10" step="1" /></div>
-
-    <div class="yellow label">Step</div>
-    <div class="yellow"><input class="controlItem" id="step" type="range" value="1" min="1" max="20" step="1" /></div>
-
-    <div class="yellow label">Exclude transparent pixels</div>
-    <div class="yellow">
-      <select class="controlItem" id="excludeTransparentPixels">
-        <option value="0">false</option>
-        <option value="1">true</option>
-      </select>
-    </div>
-
-    <div class="pink label">Include red</div>
-    <div class="pink">
-      <select class="controlItem" id="includeRed">
-        <option value="0">false</option>
-        <option value="1">true</option>
-      </select>
-    </div>
-
-    <div class="green label">Include green</div>
-    <div class="green">
-      <select class="controlItem" id="includeGreen">
-        <option value="0">false</option>
-        <option value="1">true</option>
-      </select>
-    </div>
-
-    <div class="blue label">Include blue</div>
-    <div class="blue">
-      <select class="controlItem" id="includeBlue">
-        <option value="0">false</option>
-        <option value="1">true</option>
-      </select>
-    </div>
-
-    <div class="lavender label">Include alpha</div>
-    <div class="lavender">
-      <select class="controlItem" id="includeAlpha">
-        <option value="0">false</option>
-        <option value="1">true</option>
-      </select>
-    </div>
-
-    <div class="lavender label">Process horizontal</div>
-    <div class="lavender">
-      <select class="controlItem" id="processHorizontal">
-        <option value="0">false</option>
-        <option value="1">true</option>
-      </select>
-    </div>
-
-    <div class="lavender label">Process vertical</div>
-    <div class="lavender">
-      <select class="controlItem" id="processVertical">
-        <option value="0">false</option>
-        <option value="1">true</option>
-      </select>
-    </div>
-
-    <div class="yellow label">Memoize filter</div>
+    <div class="yellow label">Memoize filters</div>
     <div class="yellow">
       <select class="controlItem" id="memoizeFilterOutput">
         <option value="0">false</option>
@@ -153,7 +80,7 @@
     <canvas 
       id="mycanvas" 
       width="400" 
-      height="400"
+      height="540"
       data-scrawl-canvas
     >
       <div class="placeholder">
@@ -180,8 +107,8 @@
 
     <ul>
       <li>Create a test Picture entity</li>
-      <li>Create a blur filter and apply it to the test entity</li>
-      <li>Update filter attributes and observe the effect they have on the entity</li>
+      <li>Create red, green, blue, cyan, magenta, yellow, notred, notgreen, notblue, grayscale, sepia and invert filters; apply them to the test entity</li>
+      <li>Update filter attribute and observe the effect it has on the entitys</li>
     </ul>
 
     <p><b>Touch test:</b> not required</p>
