+++ demo/filters-001.js
@@ -1,5 +1,5 @@
 // # Demo Filters 001
-// Parameters for: Blur, Gaussianblur filters; filter memoization
+// Filter parameters: red, green, blue, cyan, magenta, yellow, notred, notgreen, notblue, grayscale, sepia, invert
 
 // [Run code](../../demo/filters-001.html)
 import * as scrawl from '../source/scrawl.js';
@@ -20,39 +20,80 @@ const name = (n) => `${namespace}-${n}`;
 scrawl.importDomImage('.flowers');
 
 
-// Create the two blur filters
-const gaussian = scrawl.makeFilter({
+// Preparation
+const filterMethods = [
+    'red',       'green',    'blue',
+    'cyan',      'magenta',  'yellow',
+    'notred',    'notgreen', 'notblue',
+    'grayscale', 'sepia',    'invert'
+];
 
-    name: name('gaussian-blur'),
-    method: 'gaussianBlur',
-    radius: 10,
+const positions = [
+    [10, 10],  [140, 10],  [270, 10],
+    [10, 140], [140, 140], [270, 140],
+    [10, 270], [140, 270], [270, 270],
+    [10, 400], [140, 400], [270, 400],
+];
+
+const filters = [];
+
+const pictureGroup = scrawl.makeGroup({
+
+    name: name('pictures'),
+    host: canvas.getBase(),
 });
 
-const legacy = gaussian.clone({
+pictureGroup.clone({
 
-    name: name('legacy-blur'),
-    method: 'blur',
-    includeAlpha: false,
-    passes: 1,
-    step: 1,
+    name: name('labels'),
 });
 
 
-// Create the Picture entity
-const piccy = scrawl.makePicture({
+const capitalize = (item) => item[0].toUpperCase() + item.slice(1);
 
-    name: name('image'),
-    asset: 'iris',
 
-    width: '100%',
-    height: '100%',
+// Build out the filters and entitys
+filterMethods.forEach((n, index) => {
 
-    copyWidth: '100%',
-    copyHeight: '100%',
+    filters.push(scrawl.makeFilter({
 
-    method: 'fill',
+        name: name(`${n}-filter`),
+        method: n,
+    }));
 
-    filters: [name('gaussian-blur')],
+    scrawl.makePicture({
+
+        name: name(`${n}-output`),
+        group: name('pictures'),
+
+        asset: 'iris',
+
+        start: positions[index],
+        dimensions: [120, 120],
+
+        copyDimensions: ['100%', '100%'],
+
+        filters: [name(`${n}-filter`)],
+    });
+
+    scrawl.makeLabel({
+
+        name: name(`${n}-label`),
+        group: name('labels'),
+
+        text: capitalize(n),
+
+        fontString: '20px sans-serif',
+
+        fillStyle: 'white',
+        lineWidth: 4,
+
+        method: 'drawThenFill',
+
+        pivot: name(`${n}-output`),
+        lockTo: 'pivot',
+        offset: [5, 5],
+    });
 });
 
 
@@ -61,11 +102,7 @@ const piccy = scrawl.makePicture({
 const report = reportSpeed('#reportmessage', function () {
 
     return `
-Legacy:
-    Radius: ${dom.radius.value}, Step: ${dom.step.value}, Passes: ${dom.passes.value}, Opacity: ${dom.opacity.value}
-
-Gaussian:
-    Radius: ${dom.radius.value}, Opacity: ${dom.opacity.value}`;
+    Opacity: ${dom.opacity.value}`;
 });
 
 
@@ -79,155 +116,42 @@ scrawl.makeRender({
 
 
 // #### User interaction
-const enableLegacyButtons = () => {
-    dom.includeRed.removeAttribute('disabled');
-    dom.includeGreen.removeAttribute('disabled');
-    dom.includeBlue.removeAttribute('disabled');
-    dom.includeAlpha.removeAttribute('disabled');
-    dom.excludeTransparentPixels.removeAttribute('disabled');
-    dom.processHorizontal.removeAttribute('disabled');
-    dom.processVertical.removeAttribute('disabled');
-    dom.passes.removeAttribute('disabled');
-    dom.step.removeAttribute('disabled');
-};
-
-const disableLegacyButtons = () => {
-    dom.includeRed.setAttribute('disabled', '');
-    dom.includeGreen.setAttribute('disabled', '');
-    dom.includeBlue.setAttribute('disabled', '');
-    dom.includeAlpha.setAttribute('disabled', '');
-    dom.excludeTransparentPixels.setAttribute('disabled', '');
-    dom.processHorizontal.setAttribute('disabled', '');
-    dom.processVertical.setAttribute('disabled', '');
-    dom.passes.setAttribute('disabled', '');
-    dom.step.setAttribute('disabled', '');
-};
-
 // Setup form
 const dom = initializeDomInputs([
-    ['input', 'radius', '10'],
-    ['input', 'passes', '1'],
-    ['input', 'step', '1'],
     ['input', 'opacity', '1'],
-    ['select', 'blurFilter', 1],
-    ['select', 'includeRed', 1],
-    ['select', 'includeGreen', 1],
-    ['select', 'includeBlue', 1],
-    ['select', 'includeAlpha', 0],
-    ['select', 'processHorizontal', 1],
-    ['select', 'processVertical', 1],
-    ['select', 'excludeTransparentPixels', 1],
     ['select', 'memoizeFilterOutput', 0],
 ]);
 
-disableLegacyButtons();
 
-
-// Update attributes specific to the legacy filter
-scrawl.makeUpdater({
-
-    event: ['input', 'change'],
-    origin: '.controlItem',
-
-    target: legacy,
-
-    useNativeListener: true,
-    preventDefault: true,
-
-    updates: {
-
-        passes: ['passes', 'round'],
-        step: ['step', 'round'],
-
-        includeRed: ['includeRed', 'boolean'],
-        includeGreen: ['includeGreen', 'boolean'],
-        includeBlue: ['includeBlue', 'boolean'],
-        includeAlpha: ['includeAlpha', 'boolean'],
-
-        processHorizontal: ['processHorizontal', 'boolean'],
-        processVertical: ['processVertical', 'boolean'],
-        excludeTransparentPixels: ['excludeTransparentPixels', 'boolean'],
-    },
-});
-
-// Switch between gaussian and legacy blurs
-scrawl.addNativeListener(['update', 'change'], (e) => {
-
-    e.preventDefault();
-    e.stopPropagation();
-
-    const val = e.target.value;
-
-    if (val) {
-
-        // Update the filter applied to the Picture entity
-        piccy.clearFilters();
-        piccy.addFilters(name(val));
-
-        // Update UI controls
-        switch (val) {
-
-            case 'legacy-blur' :
-                enableLegacyButtons();
-                break;
-
-            case 'gaussian-blur' :
-                disableLegacyButtons();
-                break;
-        }
-    }
-
-}, dom.blurFilter);
-
-// Get the Picture entity to memoize the filter
+// Updating the Picture entitys for memoizeFilterOutput
 scrawl.makeUpdater({
 
     event: ['input', 'change'],
     origin: dom.memoizeFilterOutput,
 
-    target: piccy,
+    target: pictureGroup,
 
     useNativeListener: true,
     preventDefault: true,
 
     updates: {
+
         memoizeFilterOutput: ['memoizeFilterOutput', 'boolean'],
     },
 });
 
-// Set the filter radius (both filters)
+
+// Updating the Filter objects for opacity
 scrawl.addNativeListener(['input', 'change'], (e) => {
 
-    e.preventDefault();
-    e.stopPropagation();
-
-    const args = {
-        radius: parseFloat(e.target.value),
-    }
-
-    gaussian.set(args);
-    legacy.set(args);
-
-}, dom.radius);
-
-// Set the filter opacity (both filters)
-scrawl.addNativeListener(['input', 'change'], (e) => {
-
-    e.preventDefault();
-    e.stopPropagation();
-
-    const args = {
-        opacity: parseFloat(e.target.value),
-    }
-
-    gaussian.set(args);
-    legacy.set(args);
+    const val = parseFloat(e.target.value);
+    filters.forEach(f => f.set({ opacity: val }));
 
 }, dom.opacity);
 
 
 // #### Drag-and-Drop image loading functionality
-addImageDragAndDrop(scrawl, canvas, `#${namespace} .assets`, piccy);
+addImageDragAndDrop(scrawl, canvas, `#${namespace} .assets`, pictureGroup);
 
 
 // #### Development and testing
