+++ demo/filters-002.html
@@ -55,24 +55,16 @@
 <body>
   <div id="test-navigation">
     <a href="filters-001.html">&larr; Previous</a>
-    <a href="filters-002a.html">Next &rarr;</a>
+    <a href="filters-003.html">Next &rarr;</a>
   </div>
   <h1><a href="index.html">Scrawl-canvas v8</a> - Filters test 002</h1>
-  <h2>Parameters for: red, green, blue, cyan, magenta, yellow, notred, notgreen, notblue, grayscale, sepia, invert filters</h2>
+  <h2>Filters - cache entity output to improve render speeds</h2>
   <h3>To change image, drag-and-drop an image file over the canvas</h3>
 
   <div class="controls">
     
     <div class="lavender label">Opacity</div>
     <div class="lavender"><input class="controlItem" id="opacity" type="range" value="1" min="0" max="1" step="0.01" /></div>
-
-    <div class="yellow label">Memoize filters</div>
-    <div class="yellow">
-      <select class="controlItem" id="memoizeFilterOutput">
-        <option value="0">false</option>
-        <option value="1">true</option>
-      </select>
-    </div>
 
   </div>
 
@@ -106,10 +98,13 @@
     <h4>Test purpose</h4>
 
     <ul>
-      <li>Create a test Picture entity</li>
-      <li>Create red, green, blue, cyan, magenta, yellow, notred, notgreen, notblue, grayscale, sepia and invert filters; apply them to the test entity</li>
-      <li>Update filter attribute and observe the effect it has on the entitys</li>
+      <li>Create red, green, blue, cyan, magenta, yellow, notred, notgreen, notblue, grayscale, sepia and invert filters</li>
+      <li>Create a set of Picture entitys to display the effects of each filter on their image asset</li>
+      <li>Cache the output of each Picture entity and create new Picture entitys from those cached images</li>
+      <li>Add functionality to dynamically capture UI updates to the filters in the cached images</li>
     </ul>
+
+    <p><b>Known issue:</b> capturing the cached images takes time and requires the use of setTimeout.</p>
 
     <p><b>Touch test:</b> not required</p>
