+++ demo/filters-032.html
@@ -55,7 +55,7 @@
 <body>
   <div id="test-navigation">
     <a href="filters-031.html">&larr; Previous</a>
-    <a href="filters-101.html">Next &rarr;</a>
+    <a href="filters-033.html">Next &rarr;</a>
   </div>
   <h1><a href="index.html">Scrawl-canvas v8</a> - Filters test 032</h1>
   <h2>Parameters for: modulateOk filter</h2>
