+++ demo/filters-033.html
@@ -1,0 +1,188 @@
+<!DOCTYPE html>
+<html>
+
+<head>
+  <meta charset="utf-8">
+  <meta http-equiv="x-ua-compatible" content="ie=edge">
+  <meta name="viewport" content="width=device-width, initial-scale=1">
+  <title>Demo Filters 033</title>
+  <link rel='icon' type='image/png' href='/favicon.png'>
+  <link href="css/normalize.css" rel="stylesheet" />
+  <link href="css/tests.css" rel="stylesheet" />
+  <style>
+    /* For progressive enhancement */
+    .canvas-progressive-enhancement-wrapper {
+      display: flex;
+      justify-content: center;
+      margin: 1em 0;
+
+      & canvas {
+        box-shadow: 0px 0px 0px 1px black;
+
+        /* In no-JS environments, the image should display in place of the canvas element */
+        & .placeholder {
+
+          text-align: center;
+          width: 400px;
+          max-width: 100%;
+
+          & img {
+            width: 100%;
+          }
+        }
+        & .assets {
+          width: 0;
+          height: 0;
+          overflow: hidden;
+        }
+      }
+    }
+
+    .controls {
+      grid-template-columns: 1fr 2fr;
+
+      @media (min-width: 600px) {
+        grid-template-columns: 1fr 2fr 1fr 2fr;
+      }
+
+      @media (scripting: none) {
+        display: none;
+      }
+    }
+  </style>
+</head>
+
+<body>
+  <div id="test-navigation">
+    <a href="filters-032.html">&larr; Previous</a>
+    <a href="filters-034.html">Next &rarr;</a>
+  </div>
+  <h1><a href="index.html">Scrawl-canvas v8</a> - Filters test 033</h1>
+  <h2>Parameters for: blur filter</h2>
+  <h3>To change image, drag-and-drop an image file over the canvas</h3>
+
+  <div class="controls">
+    
+    <div class="yellow label">Radius</div>
+    <div class="yellow"><input class="controlItem" id="radius" type="range" value="10" min="0" max="60" step="1" /></div>
+
+    <div class="yellow label">Passes</div>
+    <div class="yellow"><input class="controlItem" id="passes" type="range" value="1" min="0" max="10" step="1" /></div>
+
+    <div class="yellow label">Step</div>
+    <div class="yellow"><input class="controlItem" id="step" type="range" value="1" min="1" max="20" step="1" /></div>
+
+    <div class="yellow label">Exclude transparent pixels</div>
+    <div class="yellow">
+      <select class="controlItem" id="excludeTransparentPixels">
+        <option value="0">false</option>
+        <option value="1">true</option>
+      </select>
+    </div>
+
+    <div class="pink label">Include red</div>
+    <div class="pink">
+      <select class="controlItem" id="includeRed">
+        <option value="0">false</option>
+        <option value="1">true</option>
+      </select>
+    </div>
+
+    <div class="green label">Include green</div>
+    <div class="green">
+      <select class="controlItem" id="includeGreen">
+        <option value="0">false</option>
+        <option value="1">true</option>
+      </select>
+    </div>
+
+    <div class="blue label">Include blue</div>
+    <div class="blue">
+      <select class="controlItem" id="includeBlue">
+        <option value="0">false</option>
+        <option value="1">true</option>
+      </select>
+    </div>
+
+    <div class="lavender label">Include alpha</div>
+    <div class="lavender">
+      <select class="controlItem" id="includeAlpha">
+        <option value="0">false</option>
+        <option value="1">true</option>
+      </select>
+    </div>
+
+    <div class="orange label">Process horizontal</div>
+    <div class="orange">
+      <select class="controlItem" id="processHorizontal">
+        <option value="0">false</option>
+        <option value="1">true</option>
+      </select>
+    </div>
+
+    <div class="orange label">Process vertical</div>
+    <div class="orange">
+      <select class="controlItem" id="processVertical">
+        <option value="0">false</option>
+        <option value="1">true</option>
+      </select>
+    </div>
+
+    <div class="orange label">Memoize filter</div>
+    <div class="orange">
+      <select class="controlItem" id="memoizeFilterOutput">
+        <option value="0">false</option>
+        <option value="1">true</option>
+      </select>
+    </div>
+
+    <div class="orange label">Opacity</div>
+    <div class="orange"><input class="controlItem" id="opacity" type="range" value="1" min="0" max="1" step="0.01" /></div>
+
+  </div>
+
+  <div class="canvas-progressive-enhancement-wrapper">
+    <canvas 
+      id="mycanvas" 
+      width="400" 
+      height="400"
+      data-scrawl-canvas
+    >
+      <div class="placeholder">
+        <img
+          src="thumbs/filters-033.webp"
+          alt="Screen capture image of the demo canvas"
+        />    
+      </div>
+      <div class="assets" aria-hidden="true">
+        <img
+          id="iris"
+          src="img/iris.png"
+          class="flowers"
+          alt="A purple-mauve iris flower"
+        />    
+      </div>
+    </canvas>
+  </div>
+
+  <p id="reportmessage" role="timer" aria-live="polite">To note: Javascript is not running in this browser environment. The canvas element is ignored; instead a placeholder image displays where the canvas would have appeared.</p>
+
+  <div class="testinfo">
+    <h4>Test purpose</h4>
+
+    <ul>
+      <li>Create a test Picture entity</li>
+      <li>Create a blur filter and apply it to the test entity</li>
+      <li>Update filter attributes and observe the effect they have on the entity</li>
+    </ul>
+
+    <p><b>Touch test:</b> not required</p>
+
+    <p><a href="../docs/demo/filters-033.html">Annotated code</a></p>
+  </div>
+
+  <script src="filters-033.js" type="module"></script>
+
+</body>
+
+</html>
