+++ demo/filters-034.html
@@ -5,7 +5,7 @@
   <meta charset="utf-8">
   <meta http-equiv="x-ua-compatible" content="ie=edge">
   <meta name="viewport" content="width=device-width, initial-scale=1">
-  <title>Demo Filters 002a</title>
+  <title>Demo Filters 034</title>
   <link rel='icon' type='image/png' href='/favicon.png'>
   <link href="css/normalize.css" rel="stylesheet" />
   <link href="css/tests.css" rel="stylesheet" />
@@ -54,17 +54,31 @@
 
 <body>
   <div id="test-navigation">
-    <a href="filters-002.html">&larr; Previous</a>
-    <a href="filters-003.html">Next &rarr;</a>
+    <a href="filters-033.html">&larr; Previous</a>
+    <a href="filters-101.html">Next &rarr;</a>
   </div>
-  <h1><a href="index.html">Scrawl-canvas v8</a> - Filters test 002a</h1>
-  <h2>Filters - cache output to improve render speeds</h2>
+  <h1><a href="index.html">Scrawl-canvas v8</a> - Filters test 034</h1>
+  <h2>Parameters for: gaussianblur filter</h2>
   <h3>To change image, drag-and-drop an image file over the canvas</h3>
 
   <div class="controls">
     
-    <div class="lavender label">Opacity</div>
-    <div class="lavender"><input class="controlItem" id="opacity" type="range" value="1" min="0" max="1" step="0.01" /></div>
+    <div class="yellow label">Radius</div>
+    <div class="yellow"><input class="controlItem" id="radius" type="range" value="10" min="0" max="60" step="1" /></div>
+
+    <div class="yellow">&nbsp;</div>
+    <div class="yellow">&nbsp;</div>
+
+    <div class="orange label">Opacity</div>
+    <div class="orange"><input class="controlItem" id="opacity" type="range" value="1" min="0" max="1" step="0.01" /></div>
+
+    <div class="orange label">Memoize filter</div>
+    <div class="orange">
+      <select class="controlItem" id="memoizeFilterOutput">
+        <option value="0">false</option>
+        <option value="1">true</option>
+      </select>
+    </div>
 
   </div>
 
@@ -72,12 +86,12 @@
     <canvas 
       id="mycanvas" 
       width="400" 
-      height="540"
+      height="400"
       data-scrawl-canvas
     >
       <div class="placeholder">
         <img
-          src="thumbs/filters-002.webp"
+          src="thumbs/filters-034.webp"
           alt="Screen capture image of the demo canvas"
         />    
       </div>
@@ -98,20 +112,17 @@
     <h4>Test purpose</h4>
 
     <ul>
-      <li>Create red, green, blue, cyan, magenta, yellow, notred, notgreen, notblue, grayscale, sepia and invert filters</li>
-      <li>Create a set of Picture entitys to display the effects of each filter on their image asset</li>
-      <li>Cache the output of each Picture entity and create new Picture entitys from those cached images</li>
-      <li>Add functionality to dynamically capture UI updates to the filters in the cached images</li>
+      <li>Create a test Picture entity</li>
+      <li>Create a gaussianblur filter and apply it to the test entity</li>
+      <li>Update filter attributes and observe the effect they have on the entity</li>
     </ul>
-
-    <p><b>Known issue:</b> capturing the cached images takes time and requires the use of setTimeout.</p>
 
     <p><b>Touch test:</b> not required</p>
 
-    <p><a href="../docs/demo/filters-002a.html">Annotated code</a></p>
+    <p><a href="../docs/demo/filters-034.html">Annotated code</a></p>
   </div>
 
-  <script src="filters-002a.js" type="module"></script>
+  <script src="filters-034.js" type="module"></script>
 
 </body>
