+++ demo/filters-101.html
@@ -54,7 +54,7 @@
 
 <body>
   <div id="test-navigation">
-    <a href="filters-032.html">&larr; Previous</a>
+    <a href="filters-034.html">&larr; Previous</a>
     <a href="filters-102.html">Next &rarr;</a>
   </div>
   <h1><a href="index.html">Scrawl-canvas v8</a> - Filters test 101</h1>
