+++ demo/index.html
@@ -759,19 +759,13 @@
       <div class="no-padding"><img src="thumbs/filters-001.webp" /></div>
       <div>
         <a href="filters-001.html">Filters 001</a>
-        Parameters for: Blur, Gaussianblur filters; filter memoization
-      </div>
-
-      <div class="no-padding"><img src="thumbs/filters-002.webp" /></div>
-      <div>
-        <a href="filters-002.html">Filters 002</a>
         Parameters for: red, green, blue, cyan, magenta, yellow, notred, notgreen, notblue, grayscale, sepia, invert filters
       </div>
 
       <div class="no-padding"><img src="thumbs/filters-002.webp" /></div>
       <div>
-        <a href="filters-002a.html">Filters 002a</a>
-        Cache output to improve render speeds
+        <a href="filters-002.html">Filters 002</a>
+        Filters - cache entity output to improve render speeds
       </div>
 
       <div class="no-padding"><img src="thumbs/filters-003.webp" /></div>
@@ -953,6 +947,18 @@
       <div>
         <a href="filters-032.html">Filters 032</a>
         Parameters for: modulateOk filter
+      </div>
+
+      <div class="no-padding"><img src="thumbs/filters-033.webp" /></div>
+      <div>
+        <a href="filters-033.html">Filters 033</a>
+        Parameters for: blur filter
+      </div>
+
+      <div class="no-padding"><img src="thumbs/filters-034.webp" /></div>
+      <div>
+        <a href="filters-034.html">Filters 034</a>
+        Parameters for: gaussianblur filter
       </div>
 
     </div>
