+++ docs/index.html
@@ -566,9 +566,8 @@
 
   <h4>Filters demos</h4>
   <ul>
-    <li><a href="demo/filters-001.html">Filters-001</a> Parameters for: Blur, Gaussianblur filters; filter memoization</li>
-    <li><a href="demo/filters-002.html">Filters-002</a> Parameters for: red, green, blue, cyan, magenta, yellow, notred, notgreen, notblue, grayscale, sepia, invert filters</li>
-    <li><a href="demo/filters-002a.html">Filters-002a</a> Cache output to improve render speeds</li>
+    <li><a href="demo/filters-001.html">Filters-001</a> Parameters for: red, green, blue, cyan, magenta, yellow, notred, notgreen, notblue, grayscale, sepia, invert filters</li>
+    <li><a href="demo/filters-002.html">Filters-002</a> Cache entity output to improve render speeds</li>
     <li><a href="demo/filters-003.html">Filters-003</a> Parameters for: brightness, saturation filters</li>
     <li><a href="demo/filters-004.html">Filters-004</a> Parameters for: threshold filter</li>
     <li><a href="demo/filters-005.html">Filters-005</a> Parameters for: channelstep filter</li>
@@ -599,6 +598,8 @@
     <li><a href="demo/filters-030.html">Filters-030</a> Parameters for: negative filter</li>
     <li><a href="demo/filters-031.html">Filters-031</a> Parameters for: modifyOk filter</li>
     <li><a href="demo/filters-032.html">Filters-032</a> Parameters for: modulateOk filter</li>
+    <li><a href="demo/filters-033.html">Filters-033</a> Parameters for: blur filter</li>
+    <li><a href="demo/filters-034.html">Filters-034</a> Parameters for: gaussianblur filter</li>
   </ul>
 
   <h4>Compound filters demos</h4>
